vuex added, checkpoint for course manager
