+++ components/constants.js
@@ -1,0 +1,50 @@
+export const gpaDenoms = [
+  {
+    name: '4.0',
+    minScore: 85,
+  },
+  {
+    name: '3.7',
+    minScore: 80,
+  },
+  {
+    name: '3.3',
+    minScore: 77,
+  },
+  {
+    name: '3.0',
+    minScore: 73,
+  },
+  {
+    name: '2.7',
+    minScore: 70,
+  },
+  {
+    name: '2.3',
+    minScore: 67,
+  },
+  {
+    name: '2.0',
+    minScore: 63,
+  },
+  {
+    name: '1.7',
+    minScore: 60,
+  },
+  {
+    name: '1.3',
+    minScore: 57,
+  },
+  {
+    name: '1.0',
+    minScore: 53,
+  },
+  {
+    name: '0.7',
+    minScore: 50,
+  },
+  {
+    name: '0.0',
+    minScore: 0,
+  },
+]
