+++ nuxt.config.js
@@ -65,7 +65,7 @@ export default {
    ** Plugins to load before mounting the App
    ** https://nuxtjs.org/guide/plugins
    */
-  plugins: ['@/plugins/vue-typed-js.js'],
+  plugins: ['@/plugins/vue-typed-js.js', '@/plugins/persisted-state.client.js'],
   /*
    ** Auto import components
    ** See https://nuxtjs.org/api/configuration-components
