+++ package.json
@@ -28,12 +28,14 @@
     "apexcharts": "^3.20.0",
     "bootstrap": "^4.5.2",
     "bootstrap-vue": "^2.16.0",
-    "nuxt": "^2.14.7",
-    "nuxt-mq": "^2.0.1",
-    "vue-typed-js": "^0.1.2",
     "dotenv": "^8.2.0",
     "firebase": "^7.19.1",
-    "vue-apexcharts": "^1.6.0"
+    "nuxt": "^2.14.7",
+    "nuxt-mq": "^2.0.1",
+    "vue-apexcharts": "^1.6.0",
+    "vue-typed-js": "^0.1.2",
+    "vuex": "^3.5.1",
+    "vuex-persistedstate": "^4.0.0-beta.1"
   },
   "devDependencies": {
     "@nuxtjs/color-mode": "^2.0.0",
