+++ plugins/persisted-state.client.js
@@ -1,0 +1,7 @@
+import createPersistedState from 'vuex-persistedstate'
+
+export default ({store}) => {
+  createPersistedState({
+    key: 'csscAppData',
+  })(store)
+}
